requirements.txt
django
djangorestframework
pandas
numpy
openpyxl
xlrd
